pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
